fix(networkengine): stream IDs mirror the Rust u32
Eleven streamline fields carried a Rust u32 as Go int. The width test could not pin them while they were wrong, which is why a Go-side search for narrow fields never found them. HEU-606

+++ internal/networkengine/engine_client_test.go
@@ -2297,27 +2297,27 @@ func TestEngineClient_StreamlineLifecycle(t *testing.T) {
 		UserID: user1, SponsorID: "00000000-0000-0000-0000-000000000099", Timestamp: 1001,
 	})
 	require.NoError(t, err)
-	assert.Equal(t, 1, r1.StreamID)
+	assert.Equal(t, uint32(1), r1.StreamID)
 	assert.Equal(t, 0, r1.Position)
 
 	r2, err := client.StreamlineAddMember(ctx, structure, StreamlineAddMemberRequest{
 		UserID: user2, SponsorID: user1, Timestamp: 1002,
 	})
 	require.NoError(t, err)
-	assert.Equal(t, 1, r2.StreamID)
+	assert.Equal(t, uint32(1), r2.StreamID)
 	assert.Equal(t, 1, r2.Position)
 
 	r3, err := client.StreamlineAddMember(ctx, structure, StreamlineAddMemberRequest{
 		UserID: user3, SponsorID: user1, Timestamp: 1003,
 	})
 	require.NoError(t, err)
-	assert.Equal(t, 1, r3.StreamID)
+	assert.Equal(t, uint32(1), r3.StreamID)
 
 	// Verify member info.
 	info, err := client.StreamlineGetMember(ctx, structure, user2)
 	require.NoError(t, err)
 	assert.Len(t, info.Streams, 1)
-	assert.Equal(t, 1, info.Streams[0].StreamID)
+	assert.Equal(t, uint32(1), info.Streams[0].StreamID)
 
 	// List streams.
 	streams, err := client.StreamlineListStreams(ctx, structure)
